refactor: remove cliff amount and vesting amount from VestingSchedule
