+++ contracts/airdrop-token-vesting/src/msg.rs
@@ -1,5 +1,5 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{StdResult, Timestamp, Uint128, Uint64};
+use cosmwasm_std::{Timestamp, Uint128, Uint64};
 use cw20::Denom;
 
 use crate::{
@@ -112,11 +112,9 @@ pub struct VestingData {
 #[cw_serde]
 pub enum VestingSchedule {
     LinearVestingWithCliff {
-        start_time: Uint64,      // vesting start time in second unit
-        end_time: Uint64,        // vesting end time in second unit
-        vesting_amount: Uint128, // total vesting amount
-        cliff_amount: Uint128,   // amount that will be unvested at cliff_time
-        cliff_time: Uint64,      // cliff time in second unit
+        start_time: Uint64, // vesting start time in second unit
+        end_time: Uint64,   // vesting end time in second unit
+        cliff_time: Uint64, // cliff time in second unit
     },
 }
 
@@ -150,72 +148,6 @@ impl Cliff {
 }
 
 impl VestingSchedule {
-    pub fn vested_amount(&self, block_time: u64) -> StdResult<Uint128> {
-        match self {
-            VestingSchedule::LinearVestingWithCliff {
-                start_time: _start_time,
-                end_time,
-                vesting_amount,
-                cliff_amount,
-                cliff_time,
-            } => {
-                if block_time < cliff_time.u64() {
-                    return Ok(Uint128::zero());
-                }
-
-                if block_time == cliff_time.u64() {
-                    return Ok(*cliff_amount);
-                }
-
-                if block_time >= end_time.u64() {
-                    return Ok(*vesting_amount);
-                }
-
-                let remaining_token =
-                    vesting_amount.checked_sub(*cliff_amount)?;
-                let vested_token = remaining_token
-                    .checked_mul(Uint128::from(block_time - cliff_time.u64()))?
-                    .checked_div(Uint128::from(end_time - cliff_time))?;
-
-                Ok(vested_token + cliff_amount)
-            }
-        }
-    }
-
-    ///
-    /// Validates the vesting schedule.
-    ///
-    /// - If the VestingSchedule is LinearVesting, it checks that the vesting amount is not zero.
-    /// - If the VestingSchedule is LinearVestingWithCliff, it checks:
-    ///    - that the vesting amount is not zero.
-    ///    - that the cliff amount is not zero.
-    ///    - that the cliff amount is less than or equal to the vesting amount.
-    ///
-    /// Also it calls to validate_time
-    ///
-    pub fn validate(&self, block_time: Timestamp) -> Result<(), VestingError> {
-        self.validate_time(block_time)?;
-        match self {
-            VestingSchedule::LinearVestingWithCliff {
-                start_time: _,
-                end_time: _,
-                vesting_amount,
-                cliff_time,
-                cliff_amount,
-            } => {
-                if vesting_amount.is_zero() {
-                    return Err(VestingError::ZeroVestingAmount);
-                }
-                let cliff = Cliff {
-                    amount: *cliff_amount,
-                    time: *cliff_time,
-                };
-                cliff.ok_amount(*vesting_amount)?;
-                Ok(())
-            }
-        }
-    }
-
     ///
     /// validate_time checks that the start_time is less than the end_time.
     /// additionally, if the vesting schedule is LinearVestingWithCliff, it checks that the cliff_time
@@ -224,10 +156,7 @@ impl VestingSchedule {
     /// Additionally, it the vesting schedule is LinearVestingWithCliff, it checks that the cliff_time
     /// is bigger or equal to the block_time.
     ///
-    pub fn validate_time(
-        &self,
-        block_time: Timestamp,
-    ) -> Result<(), VestingError> {
+    pub fn validate(&self, block_time: Timestamp) -> Result<(), VestingError> {
         match self {
             VestingSchedule::LinearVestingWithCliff {
                 start_time,
@@ -249,30 +178,5 @@ impl VestingSchedule {
                 Ok(())
             }
         }
-    }
-}
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use crate::contract::tests::TestResult;
-
-    #[test]
-    fn linear_vesting_with_cliff_vested_amount() -> TestResult {
-        let schedule = VestingSchedule::LinearVestingWithCliff {
-            start_time: Uint64::new(100),
-            end_time: Uint64::new(110),
-            vesting_amount: Uint128::new(1_000_000_u128),
-            cliff_amount: Uint128::new(100_000_u128),
-            cliff_time: Uint64::new(105),
-        };
-
-        assert_eq!(schedule.vested_amount(100)?, Uint128::zero());
-        assert_eq!(schedule.vested_amount(105)?, Uint128::new(100000u128)); // cliff time then the cliff amount
-        assert_eq!(schedule.vested_amount(120)?, Uint128::new(1000000u128)); // complete vesting
-        assert_eq!(schedule.vested_amount(104)?, Uint128::zero()); // before cliff time
-        assert_eq!(schedule.vested_amount(109)?, Uint128::new(820_000)); // after cliff time but before end time
-
-        Ok(())
     }
 }
