+++ contracts/airdrop-token-vesting/src/state.rs
@@ -3,7 +3,7 @@ use std::collections::HashSet;
 use cosmwasm_schema::cw_serde;
 
 use crate::msg::VestingSchedule;
-use cosmwasm_std::Uint128;
+use cosmwasm_std::{StdResult, Uint128};
 use cw_storage_plus::{Item, Map};
 
 pub const VESTING_ACCOUNTS: Map<&str, VestingAccount> =
@@ -34,6 +34,39 @@ impl Whitelist {
 pub struct VestingAccount {
     pub address: String,
     pub vesting_amount: Uint128,
+    pub cliff_amount: Uint128,
     pub vesting_schedule: VestingSchedule,
     pub claimed_amount: Uint128,
+}
+
+impl VestingAccount {
+    pub fn vested_amount(&self, block_time: u64) -> StdResult<Uint128> {
+        match self.vesting_schedule {
+            VestingSchedule::LinearVestingWithCliff {
+                start_time: _start_time,
+                end_time,
+                cliff_time,
+            } => {
+                if block_time < cliff_time.u64() {
+                    return Ok(Uint128::zero());
+                }
+
+                if block_time == cliff_time.u64() {
+                    return Ok(self.cliff_amount);
+                }
+
+                if block_time >= end_time.u64() {
+                    return Ok(self.vesting_amount);
+                }
+
+                let remaining_token =
+                    self.vesting_amount.checked_sub(self.cliff_amount)?;
+                let vested_token = remaining_token
+                    .checked_mul(Uint128::from(block_time - cliff_time.u64()))?
+                    .checked_div(Uint128::from(end_time - cliff_time))?;
+
+                Ok(vested_token + self.cliff_amount)
+            }
+        }
+    }
 }
